added new vacuum policy that uses discrete gaussain policy
