+++ bash_scripts/run_navigation_vacuum.sh
@@ -3,8 +3,8 @@ softlearning run_example_local examples.development \
     --algorithm SAC \
     --universe gym \
     --domain Locobot \
-    --task MixedNavigation-v0 \
-    --exp-name locobot-mixed-navigation-test \
+    --task NavigationVacuum-v0 \
+    --exp-name locobot-navigation-vacuum-test \
     --checkpoint-frequency 10 \
     --trial-cpus 3 \
     --trial-gpus 0 \
