+++ softlearning/algorithms/sac.py
@@ -8,6 +8,7 @@ import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
 
+from softlearning.environments.gym.spaces import *
 from softlearning.utils.gym import is_continuous_space, is_discrete_space
 from .rl_algorithm import RLAlgorithm
 
@@ -42,6 +43,10 @@ def heuristic_target_entropy(action_space):
     elif is_discrete_space(action_space):
         raise NotImplementedError(
             "TODO(hartikainen): implement for discrete spaces.")
+    elif isinstance(action_space, DiscreteBox):
+        discrete_target_entropy = -(0.9 * np.log(0.9) + 0.1 * np.log(0.1 / (action_space.num_discrete - 1)))
+        gaussian_target_entropy = -action_space.total_dimension
+        heuristic_target_entropy = discrete_target_entropy + gaussian_target_entropy
     else:
         raise NotImplementedError((type(action_space), action_space))
