+++ softlearning/environments/gym/locobot/__init__.py
@@ -2,4 +2,5 @@ from .grasping_envs import LocobotGraspingEnv, ImageLocobotGraspingEnv, ImageLoc
 # from .search_envs import LocobotSearchEnv, ImageLocobotSearchEnv
 # from .search_grasp_envs import ImageLocobotSearchGraspEnv, ImageLocobotMultiSearchGraspEnv
 # from .locobot_envs import LocobotNavigationEnv, ImageLocobotMobileGraspingEnv, LocobotMobileGraspingEnv, ImageLocobotNavigationEnv
-from .nav_grasp_envs import ImageLocobotNavigationEnv, MixedLocobotNavigationEnv, MixedLocobotNavigationReachEnv
+from .nav_envs import ImageLocobotNavigationEnv, MixedLocobotNavigationEnv, MixedLocobotNavigationReachEnv
+from .nav_grasp_envs import LocobotNavigationVacuumEnv
