+++ softlearning/replay_pools/simple_replay_pool.py
@@ -3,6 +3,8 @@ import tree
 
 from .flexible_replay_pool import FlexibleReplayPool, Field
 
+from softlearning.environments.gym.spaces import *
+
 
 def field_from_gym_space(name, space):
     if isinstance(space, spaces.Box):
@@ -12,6 +14,8 @@ def field_from_gym_space(name, space):
     elif isinstance(space, spaces.Dict):
         return tree.map_structure_with_path(
             field_from_gym_space, space.spaces)
+    elif isinstance(space, DiscreteBox):
+        return Field(name=name, dtype=space.dtype, shape=space.get_one_hot_shape())
     else:
         raise NotImplementedError(space)
 
@@ -31,14 +35,9 @@ class SimpleReplayPool(FlexibleReplayPool):
         self._action_space = action_space
 
         fields = {
-            'observations': field_from_gym_space(
-                'observations', observation_space),
-            'next_observations': field_from_gym_space(
-                'next_observations', observation_space),
-            'actions': Field(
-                name='actions',
-                dtype=action_space.dtype,
-                shape=environment.action_space.shape),
+            'observations': field_from_gym_space('observations', observation_space),
+            'next_observations': field_from_gym_space('next_observations', observation_space),
+            'actions': field_from_gym_space('actions', action_space),
             'rewards': Field(
                 name='rewards',
                 dtype='float32',
